stronger types

+++ backend/include/Component/Game/ProgressTimerIncrement.h
@@ -8,7 +8,7 @@ namespace gamecomp {
 struct ProgressTimerIncrement {
     ProgressTimer base;
 
-    double addvalue_per_percent = 1.0;
+    int addvalue_per_percent = 1.0;
     bool isendless = false;
 };
 } // namespace gamecomp
